Consolidate WiSE benchmark per PR #445 review
- Merge density / rdf / xray_sf into a single litfsi_h2o_21m benchmark
  under ml_peg/{calcs,analysis,app}/wise_electrolytes/litfsi_h2o_21m/.
- Adopt the standard ml-peg patterns in the analysis and calc scripts:
  load_models(current_models) for model discovery, @build_table for
  the metrics table, and metrics.yml for thresholds/tooltips/weights.
- Add a Janus recast of the LAMMPS+symmetrix Adastra production protocol
  in ml_peg/calcs/wise_electrolytes/md_reference/calc_md_reference.py
  (pytest-skipped reference; documents the exact MD parameters).
- Add the docs page docs/source/user_guide/benchmarks/wise_electrolytes.rst
  (and toctree entry) and wire the app docs_url to it.
- Update the parent wise_electrolytes.yml to a single benchmark weight.
