+++ docs/source/user_guide/benchmarks/index.rst
@@ -18,3 +18,4 @@ Benchmarks
     conformers
     molecular_dynamics
     defect
+    wise_electrolytes
