--- ml_peg/analysis/wise_electrolytes/density/metrics.yml
@@ -1,15 +1,0 @@
-metrics:
-  Density Error:
-    good: 0.0
-    bad: 0.17
-    unit: "g/cm3"
-    tooltip: "Absolute error in NPT density vs experimental (1.7126 g/cm3, Gilbert et al. JCED 2017)"
-    level_of_theory: Experimental (Gilbert et al., J. Chem. Eng. Data 62, 2056, 2017)
-    weight: 1
-  Density Error (%):
-    good: 0.0
-    bad: 10.0
-    unit: "%"
-    tooltip: "Percent error in NPT density vs experimental (1.7126 g/cm3, Gilbert et al. JCED 2017)"
-    level_of_theory: Experimental (Gilbert et al., J. Chem. Eng. Data 62, 2056, 2017)
-    weight: 0
