+++ ml_peg/analysis/wise_electrolytes/litfsi_h2o_21m/metrics.yml
@@ -1,0 +1,43 @@
+metrics:
+  Density Error:
+    good: 0.0
+    bad: 0.17
+    unit: "g/cm3"
+    tooltip: "Absolute error in NPT density vs experimental (1.7126 g/cm3, Gilbert et al. JCED 2017)"
+    level_of_theory: Experimental (Gilbert et al., J. Chem. Eng. Data 62, 2056, 2017)
+    weight: 1
+  Density Error (%):
+    good: 0.0
+    bad: 10.0
+    unit: "%"
+    tooltip: "Percent error in NPT density vs experimental (1.7126 g/cm3, Gilbert et al. JCED 2017)"
+    level_of_theory: Experimental (Gilbert et al., J. Chem. Eng. Data 62, 2056, 2017)
+    weight: 0
+  CN Li-O_water Error:
+    good: 0.0
+    bad: 0.5
+    unit: "-"
+    tooltip: "Absolute error in Li-O_water coordination number vs experimental reference (CN=2.0, Watanabe et al. J. Phys. Chem. B 2021, ~18.5 m, neutron diffraction with isotopic substitution)"
+    level_of_theory: Experimental (neutron diffraction, Watanabe et al. JPCB 2021)
+    weight: 1
+  CN Li-O_TFSI Error:
+    good: 0.0
+    bad: 0.5
+    unit: "-"
+    tooltip: "Absolute error in Li-O_TFSI coordination number vs experimental reference (CN=2.0, Watanabe et al. J. Phys. Chem. B 2021, ~18.5 m, neutron diffraction with isotopic substitution)"
+    level_of_theory: Experimental (neutron diffraction, Watanabe et al. JPCB 2021)
+    weight: 1
+  S(q) R-factor:
+    good: 0.0
+    bad: 0.5
+    unit: "-"
+    tooltip: "R-factor between computed and experimental X-ray S(q): sum|S_exp - S_calc| / sum|S_exp|"
+    level_of_theory: Experimental (SAXS, Zhang et al., J. Phys. Chem. B 125, 4501, 2021)
+    weight: 1
+  First Peak Position Error:
+    good: 0.0
+    bad: 0.3
+    unit: "1/A"
+    tooltip: "Absolute error in position of first S(q) peak (~1.0 A-1)"
+    level_of_theory: Experimental (SAXS, Zhang et al., J. Phys. Chem. B 125, 4501, 2021)
+    weight: 0
