--- ml_peg/analysis/wise_electrolytes/xray_sf/metrics.yml
@@ -1,15 +1,0 @@
-metrics:
-  S(q) R-factor:
-    good: 0.0
-    bad: 0.5
-    unit: "-"
-    tooltip: "R-factor between computed and experimental X-ray S(q): sum|S_exp - S_calc| / sum|S_exp|"
-    level_of_theory: Experimental (SAXS, Zhang et al., J. Phys. Chem. B 125, 4501, 2021)
-    weight: 1
-  First Peak Position Error:
-    good: 0.0
-    bad: 0.3
-    unit: "1/A"
-    tooltip: "Absolute error in position of first S(q) peak (~1.0 A-1)"
-    level_of_theory: Experimental (SAXS, Zhang et al., J. Phys. Chem. B 125, 4501, 2021)
-    weight: 0
