+++ ml_peg/app/wise_electrolytes/wise_electrolytes.yml
@@ -2,6 +2,4 @@ title: WiSE Electrolytes
 description: Structural and thermodynamic properties of 21 m LiTFSI/H2O "water-in-salt" electrolytes from molecular dynamics
 weight: 0
 benchmark_weights:
-  WiSE Density: 0.5
-  WiSE X-ray S(q): 0.3
-  WiSE Li-O RDF: 0.2
+  WiSE LiTFSI/H2O 21 m: 1.0
